+++ packages/in_app_purchase/in_app_purchase_android/test/in_app_purchase_android_platform_addition_test.dart
@@ -56,7 +56,7 @@ void main() {
       );
       final BillingResultWrapper billingResultWrapper =
           await iapAndroidPlatformAddition.consumePurchase(
-              GooglePlayPurchaseDetails.fromPurchase(dummyPurchase));
+              GooglePlayPurchaseDetails.fromPurchase(dummyPurchase).first);
 
       expect(billingResultWrapper, equals(expectedBillingResult));
     });