+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.3.0
-* **BREAKING CHANGE**: Removes `launchPriceChangeConfirmationFlow` from `InAppPurchaseAndroidPlatform` as it is deprecated by Android.
+* **BREAKING CHANGE**: Removes `launchPriceChangeConfirmationFlow` from `InAppPurchaseAndroidPlatform`. Price changes are now [handled by Google Play](https://developer.android.com/google/play/billing/subscriptions#price-change).
 * Returns both base plans and offers when `queryProductDetailsAsync` is called.
 
 ## 0.2.5+5